Upgrade libhydrogen
